+++ src/libkefir_buffer.c
@@ -13,7 +13,7 @@
 DEFINE_ERR_FUNCTIONS("buffers")
 
 __printf(3, 4)
-int buf_append(char **buf, size_t *buf_len, const char *fmt, ...)
+int buf_append(char **buf, unsigned int *buf_len, const char *fmt, ...)
 {
 	size_t offset, maxlen, reqlen;
 	va_list ap;
