+++ src/libkefir_dump.h
@@ -9,6 +9,6 @@
 #include "libkefir.h"
 
 int dump_filter_to_buf(const struct kefir_filter *filter, char **buf,
-		       size_t *buf_len, const char *prefix);
+		       unsigned int *buf_len, const char *prefix);
 
 #endif /* LIBKEFIR_DUMP_H */
